debug bullet

+++ src/GameEntity.java
@@ -4,8 +4,8 @@ import javafx.scene.image.Image;
 abstract class GameEntity
 {
     protected int i = 0, j = 0;
-    protected int x = 0;
-    protected int y = 0;
+    protected double x = 0;
+    protected double y = 0;
     Image img;
     protected double direction = 0;
